Refactor Tutor.py
