+++ open_learning_ai_tutor/intent_selector_test.py
@@ -1,79 +1,90 @@
 import pytest
 import json
-from open_learning_ai_tutor.constants import Intent, Assesment
+from open_learning_ai_tutor.constants import Intent, Assessment
 from open_learning_ai_tutor.intent_selector import get_intent
 
 
 @pytest.mark.parametrize(
-    ("assesment_selections", "previous_intent", "intents"),
+    ("assessment_selections", "previous_intent", "intents"),
     [
-        ([Assesment.IRRELEVANT_MESSAGE.value], [], [Intent.G_REFUSE]),
+        ([Assessment.IRRELEVANT_MESSAGE.value], [], [Intent.G_REFUSE]),
         (
-            [Assesment.ASKING_FOR_CONCEPTS.value],
+            [Assessment.ASKING_FOR_CONCEPTS.value],
             [],
             [Intent.S_STATE, Intent.A_CURIOSITY],
         ),
-        ([Assesment.ASKING_FOR_CALCULATION.value], [], [Intent.S_CALCULATION]),
-        ([Assesment.ASKING_FOR_DEFINITION.value], [], [Intent.S_STATE]),
-        ([Assesment.AMBIGUOUS_ANSWER.value], [], [Intent.P_ARTICULATION]),
+        ([Assessment.ASKING_FOR_CALCULATION.value], [], [Intent.S_CALCULATION]),
+        ([Assessment.ASKING_FOR_DEFINITION.value], [], [Intent.S_STATE]),
+        ([Assessment.AMBIGUOUS_ANSWER.value], [], [Intent.P_ARTICULATION]),
         (
-            [Assesment.AMBIGUOUS_ANSWER.value, Assesment.ASKING_FOR_SOLUTION.value],
+            [Assessment.AMBIGUOUS_ANSWER.value, Assessment.ASKING_FOR_SOLUTION.value],
             [],
             [Intent.P_HYPOTHESIS],
         ),
         (
             [
-                Assesment.PARTIAL_CORRECT_ANSWER.value,
-                Assesment.ASKING_FOR_CALCULATION.value,
+                Assessment.PARTIAL_CORRECT_ANSWER.value,
+                Assessment.ASKING_FOR_CALCULATION.value,
             ],
             [],
             [Intent.S_CALCULATION],
         ),
         (
             [
-                Assesment.PARTIAL_CORRECT_ANSWER.value,
-                Assesment.ASKING_FOR_CALCULATION.value,
+                Assessment.PARTIAL_CORRECT_ANSWER.value,
+                Assessment.ASKING_FOR_CALCULATION.value,
             ],
             [Intent.S_SELFCORRECTION],
             [Intent.S_CALCULATION, Intent.S_STRATEGY],
         ),
-        ([Assesment.WRONG.value], [], [Intent.S_SELFCORRECTION]),
+        ([Assessment.WRONG.value], [], [Intent.S_SELFCORRECTION]),
         (
-            [Assesment.WRONG.value],
+            [Assessment.WRONG.value],
             [Intent.S_SELFCORRECTION],
             [Intent.S_CORRECTION, Intent.S_SELFCORRECTION],
         ),
-        ([Assesment.ALGEBRAIC_ERROR.value], [], [Intent.S_SELFCORRECTION]),
+        ([Assessment.ALGEBRAIC_ERROR.value], [], [Intent.S_SELFCORRECTION]),
         (
-            [Assesment.ALGEBRAIC_ERROR.value],
+            [Assessment.ALGEBRAIC_ERROR.value],
             [Intent.S_SELFCORRECTION],
             [Intent.S_CORRECTION, Intent.S_SELFCORRECTION],
         ),
-        ([Assesment.NUMERICAL_ERROR.value], [], [Intent.S_SELFCORRECTION]),
+        ([Assessment.NUMERICAL_ERROR.value], [], [Intent.S_SELFCORRECTION]),
         (
-            [Assesment.NUMERICAL_ERROR.value],
+            [Assessment.NUMERICAL_ERROR.value],
             [Intent.S_SELFCORRECTION],
             [Intent.S_CALCULATION],
         ),
-        ([Assesment.ASKING_FOR_SOLUTION.value], [Intent.P_HYPOTHESIS], [Intent.S_HINT]),
-        ([Assesment.ASKING_FOR_SOLUTION.value], [], [Intent.P_HYPOTHESIS]),
         (
-            [Assesment.ASKING_FOR_SOLUTION.value, Assesment.INCOMPLETE_SOLUTION.value],
+            [Assessment.ASKING_FOR_SOLUTION.value],
+            [Intent.P_HYPOTHESIS],
+            [Intent.S_HINT],
+        ),
+        ([Assessment.ASKING_FOR_SOLUTION.value], [], [Intent.P_HYPOTHESIS]),
+        (
+            [
+                Assessment.ASKING_FOR_SOLUTION.value,
+                Assessment.INCOMPLETE_SOLUTION.value,
+            ],
             [],
             [Intent.S_STRATEGY, Intent.S_HINT],
         ),
-        ([Assesment.WRONG.value], [], [Intent.S_SELFCORRECTION]),
+        ([Assessment.WRONG.value], [], [Intent.S_SELFCORRECTION]),
         (
-            [Assesment.WRONG.value, Assesment.ASKING_FOR_CALCULATION.value],
+            [Assessment.WRONG.value, Assessment.ASKING_FOR_CALCULATION.value],
             [],
             [Intent.S_CALCULATION],
         ),
-        ([Assesment.INCOMPLETE_SOLUTION.value], [], [Intent.S_STRATEGY, Intent.S_HINT]),
-        ([Assesment.PARTIAL_CORRECT_ANSWER.value], [], [Intent.S_STRATEGY]),
-        ([Assesment.COMPLETE_SOLUTION.value], [], [Intent.G_GREETINGS]),
+        (
+            [Assessment.INCOMPLETE_SOLUTION.value],
+            [],
+            [Intent.S_STRATEGY, Intent.S_HINT],
+        ),
+        ([Assessment.PARTIAL_CORRECT_ANSWER.value], [], [Intent.S_STRATEGY]),
+        ([Assessment.COMPLETE_SOLUTION.value], [], [Intent.G_GREETINGS]),
     ],
 )
-def test_intent_selector(assesment_selections, previous_intent, intents):
+def test_intent_selector(assessment_selections, previous_intent, intents):
     """Test get_intent"""
-    assesments = json.dumps({"selection": "".join(assesment_selections)})
-    assert set(get_intent(assesments, previous_intent)) == set(intents)
+    assessments = json.dumps({"selection": "".join(assessment_selections)})
+    assert set(get_intent(assessments, previous_intent)) == set(intents)
