+++ open_learning_ai_tutor/prompts.py
@@ -1,5 +1,5 @@
-import open_learning_ai_tutor.utils as utils
 from open_learning_ai_tutor.constants import Intent
+from langchain_core.messages import HumanMessage, SystemMessage
 
 
 def get_problem_prompt(problem, problem_set):
@@ -26,7 +26,7 @@ Some information required to solve the problem may be in other parts of the prob
 
 ---
 
-Provide the least amount of scaffolding possible to help the student solve the problem on their own. Be succinct but acknowledge the student's progresses and right answers. Your student can only see the text you send them using your `text_student` tool, the rest of your thinking is hidden to them."""
+Provide the least amount of scaffolding possible to help the student solve the problem on their own. Be succinct but acknowledge the student's progresses and right answers. """
 
 
 intent_mapping = {
@@ -61,6 +61,63 @@ def get_intent_prompt(intents):
         for intent in intents:
             intent_prompt += intent_mapping.get(intent, "")
 
-    intent_prompt += " Your student can only see the text you send them using your `text_student` tool, the rest of your thinking is hidden to them."
-
     return intent_prompt
+
+
+def get_assessment_initial_prompt(problem, problem_set):
+    prompt = f"""A student and their tutor are working on a math problem:
+*Problem Statement*:
+{problem}
+
+This problem is in xml format and includes a solution. The problem is part of a problem set.
+
+*Problem Set*:
+
+{problem_set}
+
+Some information required to solve the problem may be in other parts of the problem set.
+
+The tutor's utterances are preceded by "Tutor:" and the student's utterances are preceded by "Student:".
+
+Analyze the last student's utterance.
+select all the feedbacks that apply from "a,b,c,d,e,f,g,h,i,j,k,l":
+
+a) The student is using or suggesting a wrong method or taking a wrong path to solve the problem
+b) The student made an error in the algebraic manipulation
+c) The student made a numerical error
+d) The student provided an intuitive or incomplete solution
+e) The student's answer is not clear or ambiguous
+f) The student correctly answered the tutor's previous question
+g) The student is explicitly asking about how to solve the problem
+h) The student is explicitly asking the tutor to state a specific theorem, definition, formula or programming command that is not the **direct answer** to the question they have to solve.
+i) The student is explicitly asking the tutor to perform a numerical calculation
+j) The student and tutor arrived at a complete solution for the entirety of the initial *Problem Statement*
+k) The student's message is *entirely* irrelevant to the problem at hand or to the material covered by the exercise.
+l) The student is asking about concepts or information related to the material covered by the problem, or is continuing such a discussion.
+
+Proceed step by step. First briefly justify your selection, then provide a string containing the selected letters.
+Answer in the following JSON format ONLY and do not output anything else:
+
+{{
+    "justification": "..",
+    "selection": ".."
+
+}}
+
+Analyze the last student's utterance.
+"""
+    return prompt
+
+
+def get_assessment_prompt(problem, problem_set, assessment_history, new_messages):
+    initial_prompt = get_assessment_initial_prompt(problem, problem_set)
+    prompt = [SystemMessage(initial_prompt)]
+
+    if len(assessment_history) > 0:
+        prompt = prompt + assessment_history
+
+    new_messages_text = ""
+    for message in new_messages:
+        new_messages_text += ' Student: "' + message.content + '"'
+    prompt.append(HumanMessage(content=new_messages_text))
+    return prompt
